Move new rkglChainLinkSelect() from library code rkgl_chain.h(,c) into sample select_test.c for suggesction
