+++ example/chain/select_test.c
@@ -1,5 +1,20 @@
 #include <roki_gl/roki_glut.h>
 
+/* This is suggstion code for rkgl_chain.c **************************************/
+
+int rkglChainLinkSelect(rkglChain *gc, rkglSelectionBuffer *sb)
+{
+  int selected_id = -1;
+
+  if( !rkglSelectionFindNearest( sb ) ) return selected_id;
+  if( rkglSelectionName(sb,0) != gc->name ||
+      rkglSelectionName(sb,1) < 0 ||
+      rkglSelectionName(sb,1) >= rkChainLinkNum(gc->chain) ) return selected_id;
+  return ( selected_id = rkglSelectionName(sb,1) );
+}
+
+/* end of suggstion code for rkgl_chain.c ***************************************/
+
 rkglCamera cam;
 rkglLight light;
